backend/scripts/requirements.txt
GitPython==3.1.45
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
